refactor(cli): hoist helper http/json/strings to prefixed sub-packages
